requirements.txt
flask==2.2.3
alembic==1.10.2
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.8
passlib==1.7.4
pyjwt==2.6.0
werkzeug==2.2.3
bcrypt==4.0.1
pydantic==1.10.7
pydantic[email]==1.10.7
cryptography==40.0.1